backend board member work
added the board members model.
added functionality to upload profile pictures.
a board member object is created whenever a user is created.
users' display in the admin panel is their full_name and if they do not have it is their username.
added name handling in csvimport
added the pillow package for image handling